+++ section10/vue3/src/composables/useCounter.js
@@ -1,0 +1,21 @@
+import { computed } from 'vue'
+
+export default function useCounter(item){ // モジュール化した関数を外に出す
+    const increment = () => {
+        item.amount++
+    }
+
+    const decrement = () => {
+        item.amount--
+    }
+
+    const totalPrice = computed(() => {
+        return item.price * item.amount
+    })
+
+    return {
+        increment,
+        decrement,
+        totalPrice
+    }
+}
